+++ frontend/__tests__/usePasswordInput.test.ts
@@ -1,8 +1,8 @@
 import { renderHook, act } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import { ERROR_MESSAGE } from '../src/pages/signup/constants/errorMessage';
-import usePasswordInput from '../src/pages/signup/hooks/usePasswordInput';
+import { ERROR_MESSAGE } from '../src/common/constants/errorMessage';
+import usePasswordWithConfirmInput from '../src/pages/signup/hooks/usePasswordWithConfirmInput';
 
 const createFakeChangeEvent = (
   value: string,
@@ -15,7 +15,7 @@ const createFakeChangeEvent = (
 describe('usePasswordInput 훅 테스트', () => {
   describe('비밀번호 길이 유효성 검사', () => {
     it('비밀번호가 5자 미만이면 에러 메시지를 반환한다', () => {
-      const { result } = renderHook(() => usePasswordInput());
+      const { result } = renderHook(() => usePasswordWithConfirmInput());
 
       act(() => {
         result.current.handlePasswordChange(createFakeChangeEvent('1234'));
@@ -27,7 +27,7 @@ describe('usePasswordInput 훅 테스트', () => {
     });
 
     it('비밀번호가 5자일 때는 유효하다', () => {
-      const { result } = renderHook(() => usePasswordInput());
+      const { result } = renderHook(() => usePasswordWithConfirmInput());
 
       act(() => {
         result.current.handlePasswordChange(createFakeChangeEvent('12345'));
@@ -37,7 +37,7 @@ describe('usePasswordInput 훅 테스트', () => {
     });
 
     it('비밀번호가 20자일 때는 유효하다', () => {
-      const { result } = renderHook(() => usePasswordInput());
+      const { result } = renderHook(() => usePasswordWithConfirmInput());
 
       act(() => {
         result.current.handlePasswordChange(
@@ -49,7 +49,7 @@ describe('usePasswordInput 훅 테스트', () => {
     });
 
     it('비밀번호가 21자면 에러 메시지를 반환한다', () => {
-      const { result } = renderHook(() => usePasswordInput());
+      const { result } = renderHook(() => usePasswordWithConfirmInput());
 
       act(() => {
         result.current.handlePasswordChange(
@@ -65,7 +65,7 @@ describe('usePasswordInput 훅 테스트', () => {
 
   describe('비밀번호 확인 불일치 검사', () => {
     it('비밀번호와 비밀번호 확인이 일치하지 않으면 에러 메시지를 반환한다', () => {
-      const { result } = renderHook(() => usePasswordInput());
+      const { result } = renderHook(() => usePasswordWithConfirmInput());
 
       act(() => {
         result.current.handlePasswordChange(
@@ -83,7 +83,7 @@ describe('usePasswordInput 훅 테스트', () => {
     });
 
     it('비밀번호와 비밀번호 확인이 일치하면 빈 문자열을 반환한다', () => {
-      const { result } = renderHook(() => usePasswordInput());
+      const { result } = renderHook(() => usePasswordWithConfirmInput());
 
       act(() => {
         result.current.handlePasswordChange(
